restructure search_index components
